% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_folds.R, R/mort_function.R
\name{make_folds}
\alias{make_folds}
\alias{cforward}
\alias{cforward_one}
\title{Forward Selection Based on C-Index/Concordance}
\usage{
make_folds(data, event_status = "mortstat", n_folds = 10, verbose = TRUE)

cforward(
  data,
  event_time = "event_time_years",
  event_status = "mortstat",
  weight_column = "WTMEC4YR_norm",
  variables = NULL,
  included_variables = NULL,
  n_folds = 10,
  seed = 1989,
  max_model_size = 50,
  verbose = TRUE,
  cfit_args = list(),
  ...
)

cforward_one(
  data,
  event_time = "event_time_years",
  event_status = "mortstat",
  weight_column = "WTMEC4YR_norm",
  variables,
  included_variables = NULL,
  verbose = TRUE,
  cfit_args = list(),
  ...
)
}
\arguments{
\item{data}{A data set to perform model selection and cross-validation.}

\item{event_status}{Character vector of length 1 with event status, passed to
\code{\link{Surv}}}

\item{n_folds}{Number of folds for Cross-validation.  If you want to run on
the full data, set to 1}

\item{verbose}{print diagnostic messages}

\item{event_time}{Character vector of length 1 with event times, passed to
\code{\link{Surv}}}

\item{weight_column}{Character vector of length 1 with weights for
model.  If no weights are available, set to \code{NULL}}

\item{variables}{Character vector of variables to perform selection.
Must be in \code{data}.}

\item{included_variables}{Character vector of variables
forced to have in the model.  Must be in \code{data}}

\item{seed}{Seed set before folds are created.}

\item{max_model_size}{maximum number of variables in the model.  Selection
will stop if reached. Note, this does not correspond to the number
of coefficients, due to categorical variables.}

\item{cfit_args}{Arguments passed to \code{\link{concordancefit}}.  If
\code{strata} is to be passed, set \code{strata_column} in this list.}

\item{...}{Additional arguments to pass to \code{\link{coxph}}}
}
\value{
A list of lists, with elements of:
\describe{
\item{full_concordance}{Concordance when fit on the full data}
\item{models}{Cox model from full data set fit, stripped of large memory
elements}
\item{cv_concordance}{Cross-validated Concordance}
\item{included_variables}{Variables included in the model, other than
those being selection upon}
}
}
\description{
Forward Selection Based on C-Index/Concordance
}
