% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_example.R
\docType{data}
\name{nhanes_example}
\alias{nhanes_example}
\title{Example Data from National Health and Nutrition Examination Survey ('NHANES')}
\format{
A \code{data.frame} with 7 columns, which are:
\describe{
\item{SEQN}{ID of participant}
\item{mortstat}{mortality status, 1-died, 0 - censored}
\item{event_time_years}{time observed}
\item{WTMEC4YR_norm}{weights normalized for survey}
\item{gender}{gender}
\item{age_years_interview}{age in years at interview}
\item{education_adult}{educational status}
}
}
\usage{
nhanes_example
}
\description{
Example Data from National Health and Nutrition Examination Survey ('NHANES')
}
\keyword{datasets}
